Make the knockout plotter calling behaviour optional by added a new entry 'option_knockout' to 'struct plotter_table' which basically is a request from that plotter backend for the content redraw routine to get called in such a way that overlapping render areas are avoided as much as possible.
It is up to the content redraw code to actually implement this
option if it is reasonably profitable.  This was and is currently done
explicitly by the html content redraw code.  On top of that the riscos
plotter code was installing the knockout plotter itself for all content
types except plaintext and SVG and this is no longer being done in this
patch.

In more detail:

- desktop/plotters.h: added struct plotter_table::option_knockout
- render/html_redraw.c(html_redraw): if the plotter backend wants the
knockout calling behaviour, install the knockout plotter which will then
call the real backend.  Also check on the return values of clg and clip
plotter calls.
- Plotter backend changes:
  -> no longer plotting in knockout mode:
     - gtk/gtk_print.c: Also removed a 2nd instance of
       "struct plotter_table plot".
     - riscos/save_draw.c
     - riscos/print.c: Also the path plotter function pointer wasn't
       filled in and this is now fixed.
     - pdf/pdf_plotters.c: Removed the flush function as this is
       optional and we only had a dummy implementation there.
  -> remaining to request knockout mode if it makes sense based on the
     content type:
     - gtk/gtk_plotters.c
     - riscos/plotters.c
     - desktop/knockout.c
- riscos/window.c: Removed what's believed an obsolete test on the
  content type which determined to additionally install the
  knockout plotter in front of the real plotter code.


git-svn-id: svn://svn.netsurf-browser.org/trunk@4823 d004b94f-0d10-0410-95b5-d60e4a4a7a9a