Complete (but quite minimal at the moment) context menu implementation for AmigaOS: Links, objects and upload file boxes have their own menu item or submenu.
Context menus can be enable/disabled and be made "sticky" with the following options:
context_menu
sticky_context_menu

Seperated clipboard related code out to clipboard.c to make it easier to maintain and
add the ability to copy URLs to the clipboard.  Copying images to the clipboard will
come later.



git-svn-id: svn://svn.netsurf-browser.org/trunk@5629 d004b94f-0d10-0410-95b5-d60e4a4a7a9a