Function to extract an integer from a css_string. Use this in font-weight parsing.
TODO: NUMBER tokens may contain [-+]? ([0-9]+ | [0-9]* '.' [0-9]+). The lexer currently doesn't cater for the [-+]? part. This is a bug (inherited from the grammar in the spec), and must be fixed.

TODO: integer_from_css_string probably wants to pass out the number of characters processed so that the client can determine if the input was valid for their context.


git-svn-id: svn://svn.netsurf-browser.org/trunk@5645 d004b94f-0d10-0410-95b5-d60e4a4a7a9a