NetSurf will now check if it is already running (ie. if ARexx port NETSURF exists), and quit and send an OPEN command to the one in memory if this is the case.
Setting files as projects of NetSurf is now possible, as is multi-select launching
from Workbench.  Multi-selects are not passed through using ARexx yet, only the first
file in the list will be opened if NetSurf is already running.

Plain text files which have no MIMETYPE tooltype are now correctly identified as
text/plain instead of text/ascii, allowing them to be opened locally.


git-svn-id: svn://svn.netsurf-browser.org/trunk@5698 d004b94f-0d10-0410-95b5-d60e4a4a7a9a