Move the declarations of iconv_initialise/iconv_finalise to a different header. This keeps the public iconv.h free of such nonsense.
Move the source for the RISC OS stubs to the distribution template tree.
We will no longer shipped compiled stubs. People are quite capable of compiling this themselves. Also take the opportunity to tidy it up a bit.

Bump the version number to 0.09

Introduce a "riscos-dist" target in Makefile-riscos.

Update various bits of documentation.


git-svn-id: svn://svn.netsurf-browser.org/trunk@5734 d004b94f-0d10-0410-95b5-d60e4a4a7a9a