When a visitor window was present on NetSurf's public screen and NetSurf was quit, the screen would fail to close.  Re-launching NetSurf would then fail to open a public screen as a public screen with the same name was already open.  The window would end up on the Workbench screen and then crash as the pointer to the screen being used was NULL.
This fixes the problem by closing the ARexx port early (this stops any new sessions
from thinking NetSurf is still running) and looping until Intuition closes the screen.
Running NetSurf will check for a public screen of name NetSurf if opening a new screen
fails and use this if present.  It will also fall back to using the Workbench screen
if both opening a new screen and locking an existing NetSurf screen fails.


git-svn-id: svn://svn.netsurf-browser.org/trunk@5886 d004b94f-0d10-0410-95b5-d60e4a4a7a9a