-Wextra is only understood by GCC 3.4.6 or later. It is a more descriptive synonym for -W.
Use -W instead which is understood by all common versions of GCC 
(i.e. GCC 2.95 supports it, which is the oldest GCC version we have to support)

Additionally, conditionalise our -Werror usage on whether we're building for 
BeOS/Haiku. On these platforms, the standard library headers result in warnings.
GCC 2 is quite happy to build this code on other platforms, so shouldn't be
the trigger for dropping -Werror (even if BeOS is the only target where GCC2 is
likely to be used).



git-svn-id: svn://svn.netsurf-browser.org/trunk@9669 d004b94f-0d10-0410-95b5-d60e4a4a7a9a