Revert previous changes. It turns out that, despite what the documentation says, when targetting RISC OS, the soname should be in the form libfoo.so.major, and *not* libfoo/so/major. However, when we create the installed symlinks and are targetting RISC OS, the symlink targets must be of the form libfoo/so/major/minor/patch, and not libfoo.so.major.minor.patch. To add extra complexity, when we're building for RISC OS, but installing in a cross-compilation environment, we want the symlinks to be valid on Unix hosts, which means libfoo.so.major.minor.patch.
The upshot of all this is that we simply cannot perform the appropriate magic in the buildsystem makefiles as they have no knowledge of whether they're installing into a cross-compilation or a target environment. Nor should they care about this, either. Therefore, the place where we get to massage the target name is in ro-ln.


git-svn-id: svn://svn.netsurf-browser.org/trunk@9791 d004b94f-0d10-0410-95b5-d60e4a4a7a9a