Simplify libwapcaplet and remove context objects.
Remove the lwc_context type from the API and ensure that all strings belong to
the one internment context.  This removes a lot of API and simplifies a lot of
function calls, however it does mean that clients of the library *MUST* be
better at reffing and unreffing strings or it'll explode.



git-svn-id: svn://svn.netsurf-browser.org/trunk@10159 d004b94f-0d10-0410-95b5-d60e4a4a7a9a